update 66
Added breadcrumbs system that is not fully operational.
New Enemy textures for region 2.
Fixed the upgrade system to correctly provide the right upgrade.
Coins system will not take out the amount for the upgrade when you upgrade at end of level. Needs looking into and will be fixed in next patch.
New level added and base work for new button on controller for the phone added.